Move Windows only functionality to Win32DPIUtil
Moving methods that are consumed only by Windows from DPIUtil to
Win32DPIUtil
