+++ bundles/org.eclipse.swt/Eclipse SWT/common/org/eclipse/swt/internal/DPIUtil.java
@@ -16,7 +16,6 @@
 package org.eclipse.swt.internal;
 
 import java.util.*;
-import java.util.function.*;
 
 import org.eclipse.swt.*;
 import org.eclipse.swt.graphics.*;
@@ -129,21 +128,6 @@ public static float pixelToPoint(float size, int zoom) {
 	return (size / scaleFactor);
 }
 
-
-/**
- * Auto-scale image with ImageData
- */
-public static ImageData scaleImageData (Device device, final ImageData imageData, int targetZoom, int currentZoom) {
-	if (imageData == null || targetZoom == currentZoom || (device != null && !device.isAutoScalable())) return imageData;
-	float scaleFactor = (float) targetZoom / (float) currentZoom;
-	return autoScaleImageData(device, imageData, scaleFactor);
-}
-
-
-public static ImageData scaleImageData (Device device, final ElementAtZoom<ImageData> elementAtZoom, int targetZoom) {
-	return scaleImageData(device, elementAtZoom.element(), targetZoom, elementAtZoom.zoom());
-}
-
 public static ImageData autoScaleImageData (Device device, final ImageData imageData, float scaleFactor) {
 	// Guards are already implemented in callers: if (deviceZoom == 100 || imageData == null || scaleFactor == 1.0f) return imageData;
 	int width = imageData.width;
@@ -177,17 +161,6 @@ public static ImageData autoScaleImageData (Device device, final ImageData image
 public static boolean isSmoothScalingEnabled() {
 	return autoScaleMethod == AutoScaleMethod.SMOOTH;
 }
-
-/**
-
- * Auto-scale ImageData to device zoom that are at given zoom factor.
- */
-public static ImageData autoScaleImageData (Device device, final ImageData imageData, int imageDataZoomFactor) {
-	if (deviceZoom == imageDataZoomFactor || imageData == null || (device != null && !device.isAutoScalable())) return imageData;
-	float scaleFactor = (float) deviceZoom / imageDataZoomFactor;
-	return autoScaleImageData(device, imageData, scaleFactor);
-}
-
 
 /**
  * Returns scaling factor from the given device zoom
@@ -241,72 +214,6 @@ public record ElementAtZoom<T>(T element, int zoom) {
 			SWT.error(SWT.ERROR_INVALID_ARGUMENT);
 		}
 	}
-}
-
-/**
- * Gets ImageData that are appropriate for the specified zoom level together
- * with the zoom level at which the image data are. If there is an image at the
- * specified zoom level, it is returned. Otherwise the next larger image at 150%
- * and 200% is returned, if existing. If none of these is found, the 100% image
- * is returned as a fallback. If provider or fallback image is not available, an
- * error is thrown.
- */
-public static ElementAtZoom<ImageData> validateAndGetImageDataAtZoom(ImageDataProvider provider, int zoom) {
-	if (provider == null) {
-		SWT.error(SWT.ERROR_NULL_ARGUMENT);
-	}
-	ElementAtZoom<ImageData> imageDataAtZoom = getElementAtZoom(z -> provider.getImageData(z), zoom);
-	if (imageDataAtZoom == null) {
-		SWT.error(SWT.ERROR_INVALID_ARGUMENT, null,
-				": ImageDataProvider [" + provider + "] returns null ImageData at 100% zoom.");
-	}
-	return imageDataAtZoom;
-}
-
-/**
- * Gets the image file path that are appropriate for the specified zoom level
- * together with the zoom level at which the image data are. If there is an
- * image at the specified zoom level, it is returned. Otherwise the next larger
- * image at 150% and 200% is returned, if existing. If none of these is found,
- * the 100% image is returned as a fallback. If provider or fallback image is
- * not available, an error is thrown.
- */
-public static ElementAtZoom<String> validateAndGetImagePathAtZoom(ImageFileNameProvider provider, int zoom) {
-	if (provider == null) {
-		SWT.error(SWT.ERROR_NULL_ARGUMENT);
-	}
-	ElementAtZoom<String> imagePathAtZoom = getElementAtZoom(z -> provider.getImagePath(z), zoom);
-	if (imagePathAtZoom == null) {
-		SWT.error(SWT.ERROR_INVALID_ARGUMENT, null,
-				": ImageFileNameProvider [" + provider + "] returns null filename at 100% zoom.");
-	}
-	return imagePathAtZoom;
-}
-
-private static <T> ElementAtZoom<T> getElementAtZoom(Function<Integer, T> elementForZoomProvider, int zoom) {
-	T dataAtOriginalZoom = elementForZoomProvider.apply(zoom);
-	if (dataAtOriginalZoom != null) {
-		return new ElementAtZoom<>(dataAtOriginalZoom, zoom);
-	}
-	if (zoom > 100 && zoom <= 150) {
-		T dataAt150Percent = elementForZoomProvider.apply(150);
-		if (dataAt150Percent != null) {
-			return new ElementAtZoom<>(dataAt150Percent, 150);
-		}
-	}
-	if (zoom > 100) {
-		T dataAt200Percent = elementForZoomProvider.apply(200);
-		if (dataAt200Percent != null) {
-			return new ElementAtZoom<>(dataAt200Percent, 200);
-		}
-	}
-	if (zoom != 100) {
-		T dataAt100Percent = elementForZoomProvider.apply(100);
-		if (dataAt100Percent != null) {
-			return new ElementAtZoom<>(dataAt100Percent, 100);
-		}
-	}
-	return null;
 }
 
 public static int getNativeDeviceZoom() {
